+++ package.json
@@ -73,6 +73,7 @@
     "eslint-plugin-import": "^2.29.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^6.1.1",
+    "fast-check": "^3.14.0",
     "jasmine": "^4.6.0",
     "json-schema-to-typescript": "^13.1.1",
     "minimist": "^1.2.8",
