+++ src/features/broker-protection/actions/captcha.js
@@ -1,12 +1,12 @@
 import { captchaCallback } from './captcha-callback.js'
-import { getElement } from './utils.js'
-import { ErrorResponse, SuccessResponse } from './types.js'
+import { getElement } from '../utils.js'
+import { ErrorResponse, SuccessResponse } from '../types.js'
 
 /**
  * Gets the captcha information to send to the backend
  *
  * @param action
- * @return {import('./types.js').ActionResponse}
+ * @return {import('../types.js').ActionResponse}
  */
 export function getCaptchaInfo (action) {
     const pageUrl = window.location.href
@@ -72,7 +72,7 @@ export function getCaptchaInfo (action) {
 * Takes the solved captcha token and injects it into the page to solve the captcha
 *
 * @param action
-* @return {import('./types.js').ActionResponse}
+* @return {import('../types.js').ActionResponse}
 */
 export function solveCaptcha (action, token) {
     const selectors = ['h-captcha-response', 'g-recaptcha-response']
