+++ src/features/broker-protection/actions/fill-form.js
@@ -1,11 +1,11 @@
-import { getElement } from './utils.js'
-import { ErrorResponse, SuccessResponse } from './types.js'
+import { getElement } from '../utils.js'
+import { ErrorResponse, SuccessResponse } from '../types.js'
 
 /**
  * Takes the solved captcha token and injects it into the page to solve the captcha
  *
  * @param action
- * @return {import('./types.js').ActionResponse}
+ * @return {import('../types.js').ActionResponse}
  */
 export function fillForm (action, userData) {
     const form = getElement(document, action.selector)
