+++ src/features/broker-protection/comparisons/constants.js
@@ -1,0 +1,141 @@
+export const nicknames = {
+    abby: ['abby', 'abigail'],
+    abigail: ['abigail', 'abby', 'abi'],
+    abraham: ['abraham', 'abe'],
+    alexander: ['alexander', 'alex'],
+    alexandra: ['alexandra', 'alex'],
+    alexis: ['alexis', 'lexi'],
+    anthony: ['anthony', 'tony'],
+    ben: ['ben', 'benjamin'],
+    benjamin: ['benjamin', 'ben'],
+    bev: ['bev', 'beverly'],
+    beverly: ['beverly', 'bev'],
+    catherine: ['catherine', 'cathy'],
+    cathy: ['cathy', 'catherine'],
+    charles: ['charles', 'charlie'],
+    charlie: ['charlie', 'charles'],
+    chris: ['chris', 'christopher'],
+    christopher: ['christopher', 'chris'],
+    clinton: ['clinton', 'clint'],
+    dan: ['dan', 'daniel', 'danny'],
+    daniel: ['daniel', 'dan', 'danny'],
+    danny: ['danny', 'dan', 'daniel'],
+    dave: ['dave', 'david'],
+    david: ['david', 'dave'],
+    don: ['don', 'donald'],
+    donald: ['donald', 'don'],
+    ed: ['ed', 'edward', 'eddie'],
+    eddie: ['eddie', 'ed', 'edward'],
+    edward: ['edward', 'eddie', 'ed'],
+    fred: ['fred', 'frederick', 'freddie'],
+    freddie: ['freddie', 'frederick', 'fred'],
+    frederick: ['frederick', 'fred', 'freddie'],
+    jacob: ['jacob', 'jake'],
+    jake: ['jake', 'jacob'],
+    james: ['james', 'jim'],
+    jeff: ['jeff', 'jeffery'],
+    jeffery: ['jeffery', 'jeff'],
+    jim: ['jim', 'james'],
+    jon: ['jon', 'jonathan'],
+    jonathan: ['jonathan', 'jon'],
+    josh: ['josh', 'joshua'],
+    joshua: ['joshua', 'josh'],
+    katherine: ['katherine', 'katie'],
+    katheryn: ['katheryn', 'katie'],
+    kim: ['kim', 'kimberly'],
+    kimberly: ['kimberly', 'kim'],
+    lexi: ['lexi', 'alexis'],
+    lucas: ['lucas', 'luke'],
+    luke: ['luke', 'lucas'],
+    matt: ['matt', 'matthew'],
+    matthew: ['matthew', 'matt'],
+    michael: ['michael', 'mike'],
+    mike: ['mike', 'michael'],
+    nate: ['nate', 'nathan', 'nathaniel'],
+    nathan: ['nathan', 'nathaniel', 'nate'],
+    nathaniel: ['nathaniel', 'nathan', 'nate'],
+    nicholas: ['nicholas', 'nick'],
+    nick: ['nick', 'nicholas'],
+    patricia: ['patricia', 'pat'],
+    ray: ['ray', 'raymond'],
+    raymond: ['raymond', 'ray'],
+    richard: ['richard', 'rich', 'rick'],
+    rob: ['rob', 'robert', 'bob'],
+    robert: ['robert', 'rob', 'bob'],
+    rus: ['rus', 'russell'],
+    russell: ['russell', 'rus'],
+    sam: ['sam', 'samuel', 'sammy'],
+    samuel: ['samuel', 'sam', 'sammy'],
+    stan: ['stan', 'stanley'],
+    stanley: ['stanley', 'stan'],
+    sue: ['sue', 'susan'],
+    susan: ['susan', 'sue'],
+    ted: ['ted', 'teddy'],
+    teddy: ['teddy', 'ted'],
+    thad: ['thad', 'thaddeus'],
+    thaddeus: ['thaddeus', 'thad'],
+    thomas: ['thomas', 'tom', 'tommy'],
+    tim: ['tim', 'timothy'],
+    timothy: ['timothy', 'tim'],
+    tom: ['tom', 'thomas', 'tommy'],
+    tommy: ['tommy', 'tom', 'thomas'],
+    tony: ['tony', 'anthony'],
+    will: ['will', 'william'],
+    william: ['william', 'will'],
+    zach: ['zach', 'zachary'],
+    zachary: ['zachary', 'zach']
+}
+
+export const states = {
+    AL: 'Alabama',
+    AK: 'Alaska',
+    AZ: 'Arizona',
+    AR: 'Arkansas',
+    CA: 'California',
+    CO: 'Colorado',
+    CT: 'Connecticut',
+    DC: 'District of Columbia',
+    DE: 'Delaware',
+    FL: 'Florida',
+    GA: 'Georgia',
+    HI: 'Hawaii',
+    ID: 'Idaho',
+    IL: 'Illinois',
+    IN: 'Indiana',
+    IA: 'Iowa',
+    KS: 'Kansas',
+    KY: 'Kentucky',
+    LA: 'Louisiana',
+    ME: 'Maine',
+    MD: 'Maryland',
+    MA: 'Massachusetts',
+    MI: 'Michigan',
+    MN: 'Minnesota',
+    MS: 'Mississippi',
+    MO: 'Missouri',
+    MT: 'Montana',
+    NE: 'Nebraska',
+    NV: 'Nevada',
+    NH: 'New Hampshire',
+    NJ: 'New Jersey',
+    NM: 'New Mexico',
+    NY: 'New York',
+    NC: 'North Carolina',
+    ND: 'North Dakota',
+    OH: 'Ohio',
+    OK: 'Oklahoma',
+    OR: 'Oregon',
+    PA: 'Pennsylvania',
+    RI: 'Rhode Island',
+    SC: 'South Carolina',
+    SD: 'South Dakota',
+    TN: 'Tennessee',
+    TX: 'Texas',
+    UT: 'Utah',
+    VT: 'Vermont',
+    VA: 'Virginia',
+    WA: 'Washington',
+    WV: 'West Virginia',
+    WI: 'Wisconsin',
+    WY: 'Wyoming'
+}
